// test/integration/test/sql-orm-client/fixtures/mn-psl-implicit/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
type DefaultLiteralValue<CodecId extends string, _Encoded> = CodecId extends keyof CodecTypes
  ? CodecTypes[CodecId]['output']
  : _Encoded;

export type FieldOutputTypes = {
  readonly public: {
    readonly _PostToTag: {
      readonly A: CodecTypes['pg/int4@1']['output'];
      readonly B: CodecTypes['pg/text@1']['output'];
    };
    readonly Post: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
    };
    readonly Tag: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly label: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly _PostToTag: {
      readonly A: CodecTypes['pg/int4@1']['input'];
      readonly B: CodecTypes['pg/text@1']['input'];
    };
    readonly Post: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
    };
    readonly Tag: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly label: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly _PostToTag: {
      readonly A: CodecTypes['pg/int4@1']['output'];
      readonly B: CodecTypes['pg/text@1']['output'];
    };
    readonly posts: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
    };
    readonly tags: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly label: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly _PostToTag: {
      readonly A: CodecTypes['pg/int4@1']['input'];
      readonly B: CodecTypes['pg/text@1']['input'];
    };
    readonly posts: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
    };
    readonly tags: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly label: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;